Deal with corrupted cache files.
When we encounter a corrupted bitcode file, parse the module.  When we
encounter a corrupted meta-file, delete the file and give the user a warning.
Unfortunately, this is the best we can easily do for metafiles because at the
point where we learn of corruption the LLVM bitcode is already merged into the
main module.