Emit an error if you try to import when there is no source file.
Prior to this, we used to load from the cache, even if a source file for the
module could not be located.  This tends to produce weird results in the case
where the CRACK_LIB_PATH isn't set correctly.  So instead, give an error in
this case unless the user has explicitly requested this behavior with the
--allow-sourceless flag.