Always escape '?' and '/' in queries.
Indicate in the comments that we do this just to be very conservative due to
the possibility of dealing with older systems.

We may want to add a flag to allow these fields to go unescaped in cases where
we trust the server they will be sent to to handle them correctly.