Add serialized generics to the meta-digest.
Public serialized generics are part of a module's interface - if the generic
changes (even in implementation) we need to rebuild everything that depends on
the generic.

This fix still isn't quite right: we should really be excluding the line number
and file names information in a generic from the meta-digest, but that's
trickier.