Fixed premature deletion of function arguments.
It was possible for a function to delete one of its own arguments and then use
it afterwards if the argument was a borrowed reference to a global or a field
variable.  Then the function could modify the value of the variable, causing
its borrowed reference to no longer be valid.

Fixed this by introducing the concept of "volatility."  A variable is volatile
if it can be modified from a function called from the current context.  An
expression is volatile if it is a borrowed reference to a volatile variable.

When calling a function, all volatile parameters (including the receiver, for
a method) are made productive by adding a call to "oper bind" (and providing
the corresponding cleanup with a call to "oper release".