Implement the @tokens and @token_str annotations.
The @token* annotations effectively allow you to define a sequence of tokens
as a constant and created a NodeList from that constant. This can be used
inside another annotation as a cleaner way to generate code.  For example, this
code uses @tokens to implement the @hello annotation which emits a statement
to print "hello world":

    void hello(CrackContext ctx) {
        tokens := @tokens { cout `hello world!` };
        tokens.expand(ctx, null);
    }

This is a big improvement over the existing forms of code generation available
to annotations because:

-   It allows generated code to be written in a form that editors and
    formatters can deal with.
-   It will allow us to avoid hacks like having to generate chunks of code in
    reverse order and add whitespace and explicit null characters at the end
    of injected strings.
-   It preserves source location information without having to add it
    expclicitly (and without the errors that this approach produces).