Added basic lambdas.
The new lambda code takes a very conservative approach.  lambdas consist of
the "lambda" keyword, followed by the return type and a parenthesized argument
list and the body of the function.  I have attempted no type inference (either
of the return type or of the argument types) because that turns out to be very
difficult under the current architecture and will likely need to be deferred
until 2.0.

Documentation changes are also not included in this change.