Time-out SSL connections that don't complete in 5 seconds.
A client can open an SSL connection and either not send any data or not send
enough data to initiate the handshake.  Time-out if the connection handshake
doesn't complete after 5 seconds.  This protects a client that has
inadvertantly connected to an SSL port and also protects the server from DOS
attacks where a malicious client consumes all available ports.