Implement first-pass command processing library.
Implement a library and annotation processor for an interface inspired by
java's JCommander and Python's argparse.  Namely, "cmd" allows you to define
argument lists and flags as functions so that their values can be accessed
from variables without the need for explicit type conversion on the part of
the programmer.