Fix nested classes derived from interface impls.
Fix the case of a nested clase within the implementor of an interface that
inherits from that implementor.  This didn't previously work because the
special "from" methods generated for an implementation were generated at the
end of the class.  If we generate them at the beginning of the class instead,
everything works just fine.