Improve support for FluidSettings.
Allow setting values and getting the default.  This should probably have been
done from the start, it was recently motivated by "jack" no longer being the
default audio driver on linux (the default seems to now be the empy string, so
maybe the setting now needs to be explicitly defined).