Fix explicitly specified terminal devices.
Though crack.exp.termui allowed you to specify a terminal device, there were a
few bugs in it.  The specified terminal device wasn't used for the "stty"
commands used to get terminal size and switch to raw/sane mode, and it also
wasn't used for user input.  This change fixes that.