Fix bug Array cleanup for interfaces
We were using _isObject() as a compile-time check against a potentially
invalid pointer which cause "oper to Object()" to be called on that pointer.
For starters, don't do that check if the pointer may not be valid, but also
check for VTableBase instead which (in addition to not having that problem)
requires less work at runtime.